Rename hawk/dove variable risk to multiple risk for clarity
